feat: Integrate deterministic vertical line input for HtmFlowRuntime
- Added a new input source, `VerticalLineInputs`, to provide a deterministic moving vertical line stimulus for the HtmFlowRuntime.
- Updated the configuration parameters in `gui_runtime.hpp` to reduce the size of the input grid and inhibition dimensions.
- Modified the input initialization in `gui_runtime.cpp` to utilize the new deterministic input method instead of random generation.
- Enhanced documentation in `main.cpp` to clarify the nature of the input used in the runtime.
